+++ CHANGELOG.md
@@ -78,6 +78,9 @@ Line wrap the file at 100 chars.                                              Th
 - Added traffic blocking during early boot, before the daemon starts, to prevent leaks in the case
   that the system service starts after a networking daemon has already configured a network
   interface.
+- When the system process is being shut down and the target state is _secured_, maintain the
+  blocking firewall rules unless systemd reports that the system isn't shutting down. This is to
+  prevent leaks that might occur during system shutdown.
 
 
 ## [android/2022.2-beta2] - 2022-09-09